Move mutator and verification function to their own files
